Setup Two-way data binding
Users should be able to edit the hero name in an <input> textbox.

The textbox should both display the hero's name property and update that property as the user types. That means data flow from the component class out to the screen and from the screen back to the class.

To automate that data flow, we setup a two-way data binding between the <input> form element and the hero.name property.