Added a click event binding
Add a click event binding to the <li> like this:

The parentheses around click tell angular to listen for the <li> element's click event. When the user clicks in the <li>, angular executes the onSelect(hero) expression.

onSelect() is a HeroesComponent method that we wrote. Angular calls it with the hero object displayed in the clicked <li>, the same hero defined previously in the *ngFor expression.